feat: Add components and update layout for patient data visualization (dashboard perawat) (#7)
* feat: Add CardPatientInfo component for displaying patient information

* chore: Add chart.js and react-chartjs-2 dependencies

This commit adds the chart.js and react-chartjs-2 dependencies to the project. These libraries will be used for creating and displaying charts in the application.

* refactor: Update Sidebar component layout and add empty div

The Sidebar component layout has been updated to include a fixed position and an additional empty div. This change improves the overall appearance and functionality of the sidebar.

* refactor: Component layout to use a minimum width of 720px

* refactor: Update CardPatientInfo prop type for increase function

* feat: Add PatientChart component for displaying patient data (Grafik Pasien)

* feat: Add FrequencyPatient chart component for displaying patient frequency data (Frekuensi Pasien Per Bulan)